requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
bootstrap4==0.1.0
celery==4.4.2
certifi==2021.10.8
chardet==3.0.4
coverage==5.3
dj-database-url==0.5.0
Django==3.0.5
django-auditlog==0.4.7
django-autocomplete-light==3.5.1
django-bootstrap-modal-forms==1.5.0
django-common-helpers==0.9.2
django-crispy-forms==1.9.0
django-cron==0.5.1
django-debug-toolbar==2.2
django-environ==0.4.5
django-filter==2.2.0
django-heroku==0.3.1
django-jsonfield==1.4.0
django-memoize==2.3.1
django-ranged-response==0.2.0
django-recaptcha==2.0.6
django-search-views==0.3.7
django-settings-export==1.2.1
django-simple-captcha==0.5.14
django-utils-six==2.0
django-widget-tweaks==1.4.8
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
elasticsearch==7.8.1
epiweeks==2.1.2
gunicorn==20.0.4
importlib-metadata==4.10.1
kombu==4.6.8
Markdown==3.3.6
numpy==1.18.2
pandas==1.0.3
Pillow==9.0.0
plotly==4.14.3
psycopg2==2.8.5
psycopg2-binary==2.8.5
PyJWT==2.3.0
python-dateutil==2.6.0
pytz==2019.3
redis==3.4.1
retrying==1.3.3
six==1.14.0
sqlparse==0.3.1
Unidecode==1.1.1
urllib3==1.26.7
vine==1.3.0
whitenoise==5.0.1
zipp==3.1.0
